Add (incomplete) cursor test for createStream
Assure that a Stream created from a string will have the file position
set to the end of the body. This matches the result of writing to a
temporary file stream resource.

The test is marked as incomplete as implementations are divided on the
file position and the PSR does not include a clear expectation.